backend/requirements.txt
fastapi
uvicorn
langchain
langchain-mistralai
python-dotenv
pydantic[email]
supabase
bcrypt
pymupdf
python-docx
python-jose[cryptography]
python-multipart
httpx
unstructured
